+++ Kernel/Modules/AgentBook.pm
@@ -30,6 +30,7 @@ sub new {
 
     # create additional objects
     $Self->{CustomerUserObject} = Kernel::System::CustomerUser->new(%Param);
+    $Self->{JSONObject}         = Kernel::System::JSON->new( %{$Self} );
 
     return $Self;
 }
@@ -38,7 +39,7 @@ sub Run {
     my ( $Self, %Param ) = @_;
 
     # get params
-    for (qw(ToCustomer CcCustomer BccCustomer)) {
+    for (qw(ToCustomer CcCustomer BccCustomer CustomerData)) {
         $Param{$_} = $Self->{ParamObject}->GetParam( Param => $_ );
     }
 
@@ -56,10 +57,13 @@ sub Run {
             User => $_,
         );
         if ( $CustomerUserData{UserEmail} ) {
-            $List{ $CustomerUserData{UserEmail} } = $CustomerUserList{$_};
+            $List{ $CustomerUserData{UserEmail} } = {
+                Email => $CustomerUserList{$_},
+                CustomerKey => $_
+            };
         }
     }
-
+    
     # build customer search autocomplete field
     $Self->{LayoutObject}->Block(
         Name => 'CustomerSearchAutoComplete',
@@ -71,13 +75,14 @@ sub Run {
         );
 
         my $Count = 1;
-        for ( reverse sort { $List{$b} cmp $List{$a} } keys %List ) {
+        for ( reverse sort { $List{$b}->{Email} cmp $List{$a}->{Email} } keys %List ) {
             $Self->{LayoutObject}->Block(
                 Name => 'Row',
                 Data => {
-                    Name  => $List{$_},
-                    Email => $_,
-                    Count => $Count,
+                Email => $List{$_}->{Email},
+                Count => $Count,
+                CustomerDataJSON =>
+                    $Self->{JSONObject}->Encode( Data => { $List{$_}->{Email} => $List{$_}->{CustomerKey} } ),
                 },
             );
             $Count++;